+++ trailscraper/cli.py
@@ -3,6 +3,7 @@ import logging
 import os
 
 import click
+import time
 
 import trailscraper
 from trailscraper import time_utils
@@ -38,8 +39,10 @@ def root_group(verbose):
               help='Start date, e.g. "2017-01-01" or "-1days". Defaults to "one day ago".')
 @click.option('--to', 'to_s', default="now", type=click.STRING,
               help='End date, e.g. "2017-01-01" or "now". Defaults to "now".')
+@click.option('--wait', default=False, is_flag=True,
+              help='Wait until events after the specified timeframe are found.')
 # pylint: disable=too-many-arguments
-def download(bucket, prefix, account_id, region, log_dir, from_s, to_s):
+def download(bucket, prefix, account_id, region, log_dir, from_s, to_s, wait):
     """Downloads CloudTrail Logs from S3."""
     log_dir = os.path.expanduser(log_dir)
 
@@ -47,6 +50,16 @@ def download(bucket, prefix, account_id, region, log_dir, from_s, to_s):
     to_date = time_utils.parse_human_readable_time(to_s)
 
     download_cloudtrail_logs(log_dir, bucket, prefix, account_id, region, from_date, to_date)
+
+    if wait:
+        last_timestamp = last_event_timestamp_in_dir(log_dir)
+        while last_timestamp <= to_date:
+            click.echo("CloudTrail logs haven't caught up to "+str(to_date)+" yet. Most recent timestamp: "+str(last_timestamp.astimezone(to_date.tzinfo))+". Trying again in 60sec.")
+
+            time.sleep(60*1)
+
+            download_cloudtrail_logs(log_dir, bucket, prefix, account_id, region, from_date, to_date)
+            last_timestamp = last_event_timestamp_in_dir(log_dir)
 
 
 @click.command("generate-policy")